+++ gridcomps/Cap/MAPL_Cap.F90
@@ -15,6 +15,7 @@ module MAPL_CapMod
    use MAPL_CapOptionsMod
    use MAPL_ServerManager
    use MAPL_ApplicationSupport
+   use, intrinsic :: iso_fortran_env, only: REAL64, INT64, OUTPUT_UNIT
    implicit none
    private
 
@@ -226,7 +227,7 @@ contains
       integer, optional, intent(out) ::rc
 
       type (ESMF_VM) :: vm
-      integer :: start_tick, stop_tick, tick_rate
+      integer(kind=INT64) :: start_tick, stop_tick, tick_rate
       integer :: status
       class(Logger), pointer :: lgr
       
@@ -267,7 +268,6 @@ contains
       end subroutine stop_timer
 
       subroutine report_throughput(rc)
-         use, intrinsic :: iso_fortran_env, only: REAL64, OUTPUT_UNIT
          integer, optional, intent(out) :: rc
 
          integer :: rank, ierror
